mail/policyd-weight: properly escape dot in regular expression
Without this escaping, mk/subst.mk sees that there are no actual changes
with the default setup.  Nevertheless, mk/scripts/subst-identity.awk does
not classify the sed command as an identity transformation because there
_might_ be the text /etc/policyd-weightXconf, and the X would match the
dot.  Therefore, subst.mk aborts the build when it is in SUBST_NOOP_OK=no
mode.