pkgtools/pkglint: update to 20.1.6
Changes since 20.1.5:

Category Makefiles must only list subdirectories in SUBDIR that actually
contain a package.  There is no need to mention the other directories.
This is the same as in the top-level Makefile, where mk/ and regress/ are
not listed.

Packages from pkgsrc-wip may have a COMMIT_MSG file with a suggested
commit message for importing the package into main pkgsrc.