pkgtools/pkgin: allow no-op SUBST block for paths
In a pkgsrc setup with default paths, the simple substitutions are all
no-ops.  The last two substitutions only apply to Makefile.in, which
means that they are no-ops for the other files.  Since the whole sed
command is not considered an identity substitution, an explicit
SUBST_NOOP_OK is needed here.

https://mail-index.netbsd.org/pkgsrc-users/2020/05/17/msg031245.html